Rename non-test files; Move standalone tests to separate file
