+++ functionaltests/standalone_test.go
@@ -24,7 +24,154 @@ import (
 	"github.com/elastic/apm-server/functionaltests/internal/ecclient"
 )
 
-func TestUpgrade_7_17_to_8_x_to_9_x__Snapshot_Standalone_to_Managed(t *testing.T) {
+func TestUpgrade_7_17_to_8_x_Snapshot_Standalone_to_Managed(t *testing.T) {
+	from := getLatestSnapshot(t, "7.17")
+	to := getLatestSnapshot(t, "8")
+	if !from.CanUpgradeTo(to.Version) {
+		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
+		return
+	}
+
+	t.Run("UpgradeFirst", func(t *testing.T) {
+		t.Parallel()
+		runner := upgradeThenManaged78Runner(from.Version, to.Version)
+		runner.Run(t)
+	})
+
+	t.Run("ManagedFirst", func(t *testing.T) {
+		t.Parallel()
+		runner := managedThenUpgrade78Runner(from.Version, to.Version)
+		runner.Run(t)
+	})
+}
+
+func TestUpgrade_7_17_to_8_x_BC_Standalone_to_Managed(t *testing.T) {
+	from := getLatestVersionOrSkip(t, "7.17")
+	to := getLatestBCOrSkip(t, "8")
+	if !from.CanUpgradeTo(to.Version) {
+		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
+		return
+	}
+
+	t.Run("UpgradeFirst", func(t *testing.T) {
+		t.Parallel()
+		runner := upgradeThenManaged78Runner(from.Version, to.Version)
+		runner.Run(t)
+	})
+
+	t.Run("ManagedFirst", func(t *testing.T) {
+		t.Parallel()
+		runner := managedThenUpgrade78Runner(from.Version, to.Version)
+		runner.Run(t)
+	})
+}
+
+func upgradeThenManaged78Runner(fromVersion, toVersion ecclient.StackVersion) testStepsRunner {
+	// Data streams in 8.x should be all ILM if upgraded to a stack < 8.15 and > 8.16.
+	checkILM := asserts.CheckDataStreamsWant{
+		Quantity:         8,
+		PreferIlm:        true,
+		DSManagedBy:      managedByILM,
+		IndicesPerDS:     1,
+		IndicesManagedBy: []string{managedByILM},
+	}
+
+	return testStepsRunner{
+		Steps: []testStep{
+			createStep{
+				DeployVersion:     fromVersion,
+				APMDeploymentMode: apmStandalone,
+			},
+			ingestV7Step{},
+			upgradeV7Step{NewVersion: toVersion},
+			ingestStep{CheckDataStream: checkILM},
+			migrateManagedStep{},
+			ingestStep{CheckDataStream: checkILM},
+			checkErrorLogsStep{
+				ESErrorLogsIgnored: esErrorLogs{
+					eventLoopShutdown,
+				},
+				APMErrorLogsIgnored: apmErrorLogs{
+					tlsHandshakeError,
+					esReturnedUnknown503,
+					refreshCache503,
+					// TODO: remove once fixed
+					populateSourcemapFetcher403,
+				},
+			},
+		},
+	}
+}
+
+func managedThenUpgrade78Runner(fromVersion, toVersion ecclient.StackVersion) testStepsRunner {
+	checkILM := asserts.CheckDataStreamIndividualWant{
+		PreferIlm:        true,
+		DSManagedBy:      managedByILM,
+		IndicesManagedBy: []string{managedByILM},
+	}
+	checkILMRollover := asserts.CheckDataStreamIndividualWant{
+		PreferIlm:        true,
+		DSManagedBy:      managedByILM,
+		IndicesManagedBy: []string{managedByILM, managedByILM},
+	}
+
+	check := map[string]asserts.CheckDataStreamIndividualWant{
+		// These data streams are created in 7.x as well, so when we ingest again
+		// in 8.x, they will be rolled-over.
+		"traces-apm-%s":                     checkILMRollover,
+		"metrics-apm.app.opbeans_python-%s": checkILMRollover,
+		"metrics-apm.internal-%s":           checkILMRollover,
+		"logs-apm.error-%s":                 checkILMRollover,
+		// These data streams are only created in 8.x, so they will only have 1 index.
+		"metrics-apm.service_destination.1m-%s": checkILM,
+		"metrics-apm.service_transaction.1m-%s": checkILM,
+		"metrics-apm.service_summary.1m-%s":     checkILM,
+		"metrics-apm.transaction.1m-%s":         checkILM,
+	}
+
+	// These data streams are created in 7.x, but not used in 8.x,
+	// so we ignore them to avoid wrong assertions.
+	ignoredDataStreams := []string{
+		"metrics-apm.app.opbeans_node-%s",
+		"metrics-apm.app.opbeans_ruby-%s",
+		"metrics-apm.app.opbeans_go-%s",
+	}
+
+	return testStepsRunner{
+		Steps: []testStep{
+			createStep{
+				DeployVersion:     fromVersion,
+				APMDeploymentMode: apmStandalone,
+			},
+			ingestV7Step{},
+			migrateManagedStep{},
+			ingestV7Step{},
+			upgradeV7Step{NewVersion: toVersion},
+			ingestStep{
+				IgnoreDataStreams:         ignoredDataStreams,
+				CheckIndividualDataStream: check,
+			},
+			checkErrorLogsStep{
+				ESErrorLogsIgnored: esErrorLogs{
+					eventLoopShutdown,
+					addIndexTemplateTracesError,
+				},
+				APMErrorLogsIgnored: apmErrorLogs{
+					tlsHandshakeError,
+					esReturnedUnknown503,
+					refreshCache503,
+					preconditionClusterInfoCtxCanceled,
+					waitServerReadyCtxCanceled,
+					grpcServerStopped,
+					// TODO: remove once fixed
+					populateSourcemapFetcher403,
+				},
+			},
+		},
+	}
+}
+
+func TestUpgrade_7_17_to_8_x_to_9_x_Snapshot_Standalone_to_Managed(t *testing.T) {
 	from7 := getLatestSnapshot(t, "7.17")
 	to8 := getLatestSnapshot(t, "8")
 	to9 := getLatestSnapshot(t, "9")
@@ -50,7 +197,7 @@ func TestUpgrade_7_17_to_8_x_to_9_x__Snapshot_Standalone_to_Managed(t *testing.T
 	})
 }
 
-func TestUpgrade_7_17_to_8_x_to_9_x__BC_Standalone_to_Managed(t *testing.T) {
+func TestUpgrade_7_17_to_8_x_to_9_x_BC_Standalone_to_Managed(t *testing.T) {
 	from7 := getLatestVersionOrSkip(t, "7.17")
 	to8 := getLatestVersionOrSkip(t, "8")
 	to9 := getLatestBCOrSkip(t, "9")
